'decimal' format is invalid in combination with TYPE_NUMBER
As per openapi specification type 'number' does not allow 'decimal' format and this breaks openapi-generator. This workaroud sets format to None if the type is not 'string' so that it is recognized as an arbitrary number by the openapi-generator, but still allows using 'decimal' format when Decimal is coerced to string. Another solution would be setting FORMAT_DECIMAL to None in openapi.py but I think it would be less flexible and might break existiinig implementations.