Switch to using DRF JsonOpenAPIRenderer for openapi format
This required defining a `with_validators` method that is a no-op
in order to still work with how "validators" are injected into
renderers at the time that the schema view is built.

All tests are still passing after this was swapped out, so it appears
as though validators were not tested with this view.