Swap `SwaggerYAMLRenderer` with DRF one
This still does custom rendering of the YAML because we are using
custom classes that aren't natively compatible with the YAML dumper.
The eventual goal would be to do this without a custom renderer, or
to allow DRF to have a custom dumper for YAML like it does for JSON.

This also needs the `with_validators` override because the schema
view expects this to always be defined. Once again, it looks like we
don't have any actual tests which use the validators.