Special case UI rendering instead of everything else
We were special casing the URL patterns for the generator for the UI
views, but instead of looking for the UI views we were looking for
any of the schema renderers. Since we are switching the base classes
for the schema renderers, we should be instead looking for the common
UI rendering subclass.