Move validation to SchemaView
This brings us further away from the DRF-provided SchemaView, but
hopefully we can get a hook in DRF that would allow for this
validation to be provided. Right now DRF throws an error if the
generated schema is empty, so this is basically another check that
would occur around that time.

The logger for validations has changed as a result of it now
occurring in a different file.