Replace is_callable_method with is_simple_callable
This method is a more complex version of the `is_simple_callable`
method used by DRF internally. Given that this method shows no signs
of being removed soon, and it appears to be slightly simpler than
the one used within this project, it makes sense to just use that
one instead.