Declare Django 3.2 support in README, classifiers
This commit will make sure that `Django :: 3.2` will show up in the
classifiers list on PyPI: https://pypi.org/project/drf-yasg/

(The magic happens because we parse the README to get supported Django):
https://github.com/DavidCain/drf-yasg/blob/ee29412d3cdb311/setup.py#L36

3.2 support should already exist
================================
A closed pull request, https://github.com/axnsan12/drf-yasg/pull/735,
noted support for Django 3.2 in both `tox.ini`, and the README.

That PR was closed in favor of
https://github.com/axnsan12/drf-yasg/pull/741, which edited `tox.ini`
and switched to GitHub Actions.

https://github.com/axnsan12/drf-yasg/pull/735#issuecomment-924298555

This project has been testing on Django 3.2 for a long time (about a
year). I think we can declare it supported it the README!